impl Default for Level
if you have a constructor without any argument for a given type, implementing
Default is a rust idiom. It allows the user to easily find such a function.

If the normal constructor takes arguments, it is an associated function named new

Also refactor a bit a mmatch, using the unimplemented!() macro to handle the
not full pattern.